+++ CHANGELOG.md
@@ -8,7 +8,7 @@ All notable changes to this project will be documented in this file.
 - Add `COMMAND HELP` command
 - Add `SELECT` command
 - Support no arguments for `COMMAND` command
-- Support authentication for `HELLO` command
+- Support `AUTH` and `SETNAME` argument for `HELLO` command
 
 ## [v0.1.0] - 2023-02-26